Add property collections
A property is a set label in org files, but we want to have a list of
related properties, we call a collection (like all the items in a
backpack, or something). We want to be able to add and query for this
list based on some key value, like name.