Adding a new "Odds" function for not-so-50/50 rolls
Created a `rpgtk-odds` function for rolling 50/50 coin tosses, but
with shifts based on how "likely" the question would be yes (as well
as a Chaos modifier, from Mythic GM Emulator). Returns six results of
the style, Yes, and ... or No, but ... This is based on Howard's essay,
https://howardism.org/RPG/programming-yes-but.html

This also makes the `rpgtk` create a hydra of functions. While this
seems useful atm, it does depend on Hydra, so this is completely
optional.

This also includes all the regular polyhedral dice rolls, as well as a
simple Fate dice roller.

And lastly, this removes all of the complicated loading, and assumes
that the user will extend the load-path.