Fix bugs in loading by pulling hydra into own file
Having a user interface depending on hydra is nice, but that should be
its own file. This helps making the toolkit a toolkit. And it helps
with file loading.

Again, depending on `load-path` for the requires.

Also, fixed a number of bad function references (due to namespacing)
in the odds file (taken from my essay).