Create code for randomly drawing cards from a deck.
This includes the ability to make your own card decks with lists of
numbers and suits (and optionally trump cards, e.g. jokers).

Function exists for drawing a random card and returning it, or
shuffling a deck and drawing through the deck until the deck is empty,
e.g. nil.