Create code for randomly drawing cards from a deck (Part 2)
This includes the ability to make your own card decks with lists of
numbers and suits (and optionally trump cards, e.g. jokers).

Function exists for drawing a random card and returning it, or
shuffling a deck and drawing through the deck until the deck is empty,
e.g. nil.

Includes the code this time ... and a note in the README.